dashboard phase 1 done to-do in phase 2 multi grid position clear the data after create the view success and error callback new option for last X minutes/hours/days....
Mobile to-To
Work with jQuery mobile

+++ app/assets/javascripts/dashboard.js
@@ -158,7 +158,9 @@ function db_view_create_callback(data){
 }
 
 function db_view_delete(id){
+    if (confirm('You are about to delete a view from your dashboard,are you sure?')){
     ifepm.dashboard.delete_item(id,{success:db_view_delete_callback})
+    }
 }
 
 function db_view_delete_callback(data){
